+++ changelog.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
     * Requests authentication tests.
     * New borrowing validation tests.
     * End borrowing validation tests.
+    * Add inventory item validation tests.
 * Models factories
     * Users.
     * Inventory items.